CardContainer components
CardContainer already existed, but was inside of the original index.tsx - it might be best for this to be moved to a root level folder for components at some point.

"Edit" and "ListItem" card containers were in the index.tsx as their raw forms, but have been moved to their own components to make code cleaner.

Next step from here is to trigger a query when changing the "sort" of multiple "processes"