onDragEnd basic functionality
NOT COMPLETE
Current issue is that the "currentIdsArrayWithFetchedItems" doesn't have the correct ordering.  Currently, this is not returning the "sort" (which will eventually be the "orderBy" variable) and that's required in order to make sure it's ordered properly.

After fixing that, we also need to figure out how to set the items properly, because I don't think the "setItems" function works that way.  Though, maybe after getting the sorting working properly, this will work as intended as well.